+++ server/application/services/ConversationService.py
@@ -1,35 +1,11 @@
-import enum;
 import logging;
 
-from typing import Any, Optional;
+from typing import Optional;
 from mongodb.MongoDB import MongoDB;
 
-class ServiceStatus(enum.Enum):
-  SUCCESS = "success"
-  ERROR   = "error"
+from .dto import ServiceResult;
 
-  def equals(self, to: "ServiceStatus") -> bool:
-    return self.value == to.value
-
-class ServiceResult():
-  status  : ServiceStatus
-  message : str
-  data    : Any
-
-  def __init__(self, status: ServiceStatus, message: str, data: Any) -> None:
-    self.status  = status
-    self.message = message
-    self.data    = data
-
-  @classmethod
-  def success(cls, data: Any, message: str = "success"):
-    return cls(status=ServiceStatus.SUCCESS, message=message, data=data)
-
-  @classmethod
-  def error(cls, message: str = "error"):
-    return cls(status=ServiceStatus.ERROR, message=message, data=None)
-
-class Service():
+class ConversationService():
   log: logging.Logger
   database: MongoDB
